--- code/src/main/java/com/project/accountbook/card/model/MyCardDTO.java
@@ -1,137 +1,0 @@
-package com.project.accountbook.card.model;
-
-//tblMyCard
-public class MyCardDTO {
-	
-	private String cardNumber;
-	private String alias;
-	private String validity;
-	private String idMember;
-	private String seqCardInformation;
-	
-	private String ciName;
-	private String explanation;
-	private int annualFee;
-	private String overseasUse;
-	private String cardCompany;
-	private String fileLink;
-	private int seqCardType;
-	
-	private int seqReasonsChangeList;
-	private int seqMyCard;
-	
-	
-	public String getCardNumber() {
-		return cardNumber;
-	}
-	
-	public void setCardNumber(String cardNumber) {
-		this.cardNumber = cardNumber;
-	}
-	
-	public String getAlias() {
-		return alias;
-	}
-	
-	public void setAlias(String alias) {
-		this.alias = alias;
-	}
-	
-	public String getValidity() {
-		
-		return validity;
-	}
-	public void setValidity(String validity) {
-		this.validity = validity;
-	}
-	
-	public String getIdMember() {
-		return idMember;
-	}
-	
-	public void setIdMember(String idMember) {
-		this.idMember = idMember;
-	}
-	
-	public String getSeqCardInformation() {
-		return seqCardInformation;
-	}
-	
-	public void setSeqCardInformation(String seqCardInformation) {
-		this.seqCardInformation = seqCardInformation;
-	}
-	
-	public String getCiName() {
-		return ciName;
-	}
-	
-	public void setCiName(String ciName) {
-		this.ciName = ciName;
-	}
-	
-	public String getExplanation() {
-		return explanation;
-	}
-	
-	public void setExplanation(String explanation) {
-		this.explanation = explanation;
-	}
-	
-	public int getAnnualFee() {
-		return annualFee;
-	}
-	
-	public void setAnnualFee(int annualFee) {
-		this.annualFee = annualFee;
-	}
-	
-	public String getOverseasUse() {
-		return overseasUse;
-	}
-	
-	public void setOverseasUse(String overseasUse) {
-		this.overseasUse = overseasUse;
-	}
-	
-	public String getCardCompany() {
-		return cardCompany;
-	}
-	
-	public void setCardCompany(String cardCompany) {
-		this.cardCompany = cardCompany;
-	}
-	
-	public String getFileLink() {
-		return fileLink;
-	}
-	
-	public void setFileLink(String fileLink) {
-		this.fileLink = fileLink;
-	}
-	
-	public int getSeqCardType() {
-		return seqCardType;
-	}
-	
-	public void setSeqCardType(int seqCardType) {
-		this.seqCardType = seqCardType;
-	}
-	
-	public int getSeqReasonsChangeList() {
-		return seqReasonsChangeList;
-	}
-	
-	public void setSeqReasonsChangeList(int seqReasonsChangeList) {
-		this.seqReasonsChangeList = seqReasonsChangeList;
-	}
-	
-	public int getSeqMyCard() {
-		return seqMyCard;
-	}
-	
-	public void setSeqMyCard(int seqMyCard) {
-		this.seqMyCard = seqMyCard;
-	}
-	
-	
-}
