+++ code/src/main/java/com/project/accountbook/user/member/repository/MemberInfoDAO.java
@@ -5,7 +5,9 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.Statement;
+import java.util.ArrayList;
 
+import com.project.accountbook.user.member.model.MemberInfoDTO;
 import com.project.accountbook.util.DBUtil;
 
 
@@ -17,7 +19,132 @@ public class MemberInfoDAO {
 
 		public MemberInfoDAO() {
 			this.conn = DBUtil.open("192.168.10.47", "jspProject", "java1234");
-}
+		}
 
+		public int addMycard(MemberInfoDTO dto) {
+			
+			try {
 
+				String sql = "insert into tblMyCard (seq, cardNumber, alias, validity, idMember, seqCardInformation) values (seqMyCard.nextVal, ?, ?, ?, ?)";
+				pstat = conn.prepareStatement(sql);
+				pstat.setString(1, dto.getCardNumber());
+				pstat.setString(2, dto.getAlias());
+				pstat.setString(3, dto.getValidity());
+				pstat.setString(4, dto.getMcIdMember());
+				pstat.setString(5, dto.getSeqCardInformation());
+
+				pstat.executeUpdate();
+
+			} catch (Exception e) {
+				System.out.println("CardDAO.addMyCard");
+				e.printStackTrace();
+			}		
+			
+			return 0;
+		}
+		
+		
+		public int delMyCard(MemberInfoDTO dto) {
+			
+			try {
+
+				String sql = "update tblMyCard set cardNumber = 'xxxxxxxxxxxxxxxx' where cardnumber = ?";
+
+				pstat = conn.prepareStatement(sql);
+				pstat.setString(1, dto.getCardNumber());
+
+				pstat.executeUpdate();
+
+			} catch (Exception e) {
+				System.out.println("CardDAO.delMyCard");
+				e.printStackTrace();
+			}
+			return 0;
+		}
+		
+		public int updateMyCard(MemberInfoDTO dto) {
+			
+			try {
+
+				String sql = "update tblMyCard set alias = ? where cardnumber = ?";
+
+				pstat = conn.prepareStatement(sql);
+				pstat.setString(1, dto.getAlias());
+				pstat.setString(2, dto.getCardNumber());
+
+				pstat.executeUpdate();
+
+			} catch (Exception e) {
+				System.out.println("CardDAO.updateMyCard");
+				e.printStackTrace();
+			}
+			
+			return 0;
+		}
+
+		public MemberInfoDTO getMyCard(String cardNumber) {
+			
+			try {
+				
+				String sql = "select * from tblMyCard where cardnumber = ?";
+				
+				pstat = conn.prepareStatement(sql);
+				pstat.setString(1, cardNumber);
+				
+				rs = pstat.executeQuery(sql);
+				
+				if (rs.next()) {
+					
+					MemberInfoDTO dto = new MemberInfoDTO();
+					
+					dto.setAlias(rs.getString("alias"));
+					dto.setValidity(rs.getString("validity"));
+					dto.setMcIdMember(rs.getString("McIdMember"));
+					dto.setSeqCardInformation(rs.getString("seqCardInformation"));
+					
+					return dto;				
+				}	
+				
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+			
+			return null;
+		}
+		
+		
+		public ArrayList<MemberInfoDTO> listMyCard(String McIdMember){
+			
+			try {
+				
+				String sql = "select * from tblmycard where idMember = ?";
+				
+				pstat = conn.prepareStatement(sql);
+				pstat.setString(1, McIdMember);
+				
+				rs = pstat.executeQuery();
+				
+				ArrayList<MemberInfoDTO> list = new ArrayList<MemberInfoDTO>();
+				
+				while (rs.next()) {
+					
+					MemberInfoDTO dto = new MemberInfoDTO();
+
+					
+					dto.setCardNumber(rs.getString("cardNumber"));
+					dto.setAlias(rs.getString("alias"));
+					dto.setValidity(rs.getString("validity"));
+					dto.setMcIdMember(rs.getString("mcIdMember"));
+					dto.setExplanation(rs.getString("seqCardInformation"));
+					
+					list.add(dto);				
+				}	
+				return list;
+				
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+			return null;
+			
+		}
 }
